Revert to using AssemblyConvention namespace for MustIncludeAllMatchingFilesInFolder test.
