--- Conventional/Convention.Project.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using Conventional.Conventions.Projects;
-
-namespace Conventional
-{
-    public static partial class Convention
-    {
-        public static MustIncludeAllMatchingFilesInFolderConventionSpecification MustIncludeAllMatchingFilesInFolder(string filePattern)
-        {
-            return new MustIncludeAllMatchingFilesInFolderConventionSpecification(filePattern);
-        }
-    }
-}
